made separate folders for javascripts files
