+++ scripts/main.js
@@ -1,7 +1,7 @@
-import { fishList } from "./FishList.js"
+import { fishList } from "./Fish/FishList.js"
 
 fishList()
 
-import { locationList } from "./LocationList.js"
+import { locationList } from "./Locations/LocationList.js"
 
 locationList()
